Wait for Max PIN length before generating credentials
This change fixes a bug where if the handler for `RequirePINforRemoteOperation`
runs before the `maxCodeLength` attribute of the `lockCodes` capability is set,
a default pin length of 4 would be used, which may be shorter than the minimum
pin length of the device specified by the `MinPINCodeLength` attribute.
Therefore, the read was removed from `init` and `added`, and was moved to the
handler for `MaxPINCodeLength` to guarantee that the max pin length is known
before attempting to generate the COTA credential.